+++ notendb/insert_komponist.php
@@ -16,7 +16,7 @@ include('head.php');
   <tr>    
     <label>
     <td class="eingabe">Nachname:</td>  
-    <td class="eingabe"><input type="text" name="Nachname" size="45" maxlength="80" required="required" autofocus="autofocus"></td>
+    <td class="eingabe"><input type="text" name="Nachname" size="45" maxlength="80" required="required" autofocus="autofocus" oninput="changeBackgroundColor(this)"></td>
      </label>
    </tr>