+++ notendb/insert_lookup.php
@@ -35,7 +35,7 @@ if (isset($_POST["LookupTypeID"])) {
   <tr>    
     <label>
     <td class="eingabe">Name: </td>  
-    <td class="eingabe"><input type="text" name="Name" size="100" maxlength="80" required="required" autofocus="autofocus"></td>
+    <td class="eingabe"><input type="text" name="Name" size="100" maxlength="80" required="required" autofocus="autofocus" oninput="changeBackgroundColor(this)"></td>
      </label>
   </tr> 
   <tr>